+++ tests/tests.py
@@ -1,10 +1,13 @@
 from datetime import datetime
 
+import django
+from django.contrib.auth.backends import ModelBackend
 from django.contrib.auth.models import User
 from django.http import HttpRequest
 from django.template.loader import render_to_string
 from django.test import TestCase
 
+from django_extra_tools.auth.backends import SuperUserAuthenticateMixin
 from django_extra_tools.db import pg_version
 from django_extra_tools.db.models.aggregates import First, Last, Median, \
     StringAgg
@@ -215,3 +218,41 @@ class TimestampableTestCase(TestCase):
         self.assertIsInstance(self.obj.deleted_by, User)
         self.assertEqual(self.obj.deleted_by, user)
         self.assertIsInstance(self.obj.deleted_at, datetime)
+
+
+class SuperUserAuthenticateMixinTestCase(TestCase):
+    fixtures = ['superuser-authenticate-mixin']
+
+    class MyBackend(SuperUserAuthenticateMixin, ModelBackend):
+        pass
+
+    backend = MyBackend()
+
+    def authenticate(self, username, password):
+        if django.VERSION[:2] < (1, 11):
+            return self.backend.authenticate(username, password)
+        else:
+            return self.backend.authenticate(None, username, password)
+
+    def test_user(self):
+        """Test authenticate as user"""
+        user = self.authenticate('user', 'test')
+        self.assertIsInstance(user, User)
+        self.assertEqual(user.username, 'user')
+
+    def test_superuser(self):
+        """Test authenticate as superuser"""
+        user = self.authenticate('superuser', 'test')
+        self.assertIsInstance(user, User)
+        self.assertEqual(user.username, 'superuser')
+
+    def test_user_through_superuser(self):
+        """Test authenticate as user through superuser username and password"""
+        user = self.authenticate('superuser:user', 'test')
+        self.assertIsInstance(user, User)
+        self.assertEqual(user.username, 'user')
+
+    def test_unknown(self):
+        """Test authenticate as invalid user"""
+        user = self.authenticate('user123', 'test')
+        self.assertIsNone(user)