+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## master
 
-* Add `SuperUserAuthenticateMixin`
+* Add `ThroughSuperuserModelBackend`
 
 ## 0.3.0